fix bugs

+++ src/main/java/com/driver/StudentController.java
@@ -26,6 +26,8 @@ public class StudentController {
     StudentService studentService;
     @PostMapping("/add-student")
     public ResponseEntity<String> addStudent(@RequestBody Student student){
+        if (student == null) return new ResponseEntity<>("No student added", HttpStatus.BAD_REQUEST);
+
         String response = studentService.addStudent(student);
 
         return new ResponseEntity<>(response, HttpStatus.CREATED);
@@ -33,54 +35,85 @@ public class StudentController {
 
     @PostMapping("/add-teacher")
     public ResponseEntity<String> addTeacher(@RequestBody Teacher teacher){
-        String reponse = studentService.addTeacher(teacher);
+        if(teacher == null) return new ResponseEntity<>("No teacher added", HttpStatus.BAD_REQUEST);
 
-        return new ResponseEntity<>(reponse, HttpStatus.CREATED);
+        String response = studentService.addTeacher(teacher);
+
+        return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
 
     @PutMapping("/add-student-teacher-pair")
     public ResponseEntity<String> addStudentTeacherPair(@RequestParam String student, @RequestParam String teacher){
+        if (student.equals("") || teacher.equals("")) return new ResponseEntity<>("Provide both student and teacher", HttpStatus.BAD_REQUEST);
+
         String response = studentService.addStudentTeacherPair(student, teacher);
 
-        return new ResponseEntity<>(response, HttpStatus.NO_CONTENT);
+        if (response.equals("")) return new ResponseEntity<>("Pair not created", HttpStatus.BAD_REQUEST);
+
+        return new ResponseEntity<>(response, HttpStatus.CREATED);
     }
 
     @GetMapping("/get-student-by-name/{name}")
     public ResponseEntity<Student> getStudentByName(@PathVariable String name){
+        if (name.equals("")) return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+
         Student student = studentService.getStudentByName(name); // Assign student by calling service layer method
 
-        return new ResponseEntity<>(student, HttpStatus.CREATED);
+        if (student == null) return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity<>(student, HttpStatus.OK);
     }
 
     @GetMapping("/get-teacher-by-name/{name}")
     public ResponseEntity<Teacher> getTeacherByName(@PathVariable String name){
+
+        if (name.equals("")) return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+
         Teacher teacher = studentService.getTeacherByName(name); // Assign student by calling service layer method
 
-        return new ResponseEntity<>(teacher, HttpStatus.CREATED);
+        if (teacher == null) return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity<>(teacher, HttpStatus.OK);
     }
 
     @GetMapping("/get-students-by-teacher-name/{teacher}")
     public ResponseEntity<List<String>> getStudentsByTeacherName(@PathVariable String teacher){
+
+        if (teacher.equals("")) return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+
         List<String> students = studentService.getStudentsByTeacherName(teacher); // Assign list of student by calling service layer method
 
-        return new ResponseEntity<>(students, HttpStatus.CREATED);
+        if (students == null) return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity<>(students, HttpStatus.OK);
     }
 
     @GetMapping("/get-all-students")
     public ResponseEntity<List<String>> getAllStudents(){
         List<String> students = studentService.getAllStudents(); // Assign list of student by calling service layer method
 
-        return new ResponseEntity<>(students, HttpStatus.CREATED);
+        if (students == null) return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity<>(students, HttpStatus.OK);
     }
 
     @DeleteMapping("/delete-teacher-by-name")
     public ResponseEntity<String> deleteTeacherByName(@RequestParam String teacher){
+
+        if (teacher.equals("")) return new ResponseEntity<>("Enter teacher name", HttpStatus.BAD_REQUEST);
+
         String response = studentService.deleteTeacherByName(teacher);
-        return new ResponseEntity<>(teacher + " " +  response, HttpStatus.CREATED);
+
+        if (response.equals("")) return new ResponseEntity<>("No teacher found", HttpStatus.NOT_FOUND);
+        
+        return new ResponseEntity<>(teacher + " " +  response, HttpStatus.OK);
     }
     @DeleteMapping("/delete-all-teachers")
     public ResponseEntity<String> deleteAllTeachers(){
         String response = studentService.deleteAllTeachers();
-        return new ResponseEntity<>(response, HttpStatus.CREATED);
+
+        if (response.equals("")) return new ResponseEntity<>("No teacher found", HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 }
