+++ src/main/java/com/driver/StudentTeacherRepository.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Repository;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -22,9 +23,14 @@ public class StudentTeacherRepository {
 
     public String addStudentTeacherPair(String student, String teacher) {
         if(studentRepository.studentDb.containsKey(student) && teacherRepository.teacherDb.containsKey(teacher)) {
-            List<String> students = studentTeacherPairDb.getOrDefault(teacher, null);
-            students.add(student);
-            studentTeacherPairDb.put(teacher, students);
+
+            if (studentTeacherPairDb.containsKey(teacher)) {
+                List<String> oldList = studentTeacherPairDb.get(teacher);
+                oldList.add(student);
+            } else {
+                List<String> newList = new ArrayList<>();
+                newList.add(student);
+            }
 
             return "New student-teacher pair added successfully";
         }
@@ -33,13 +39,18 @@ public class StudentTeacherRepository {
     }
 
     public List<String> getStudentsByTeacherName(String teacher) {
+        if (!studentTeacherPairDb.containsKey(teacher)) return new ArrayList<>();
+
         return studentTeacherPairDb.get(teacher).stream().collect(Collectors.toList());
     }
 
-    public String deleteStudentTeacherPair(String name) {
-        List<String> students = studentTeacherPairDb.get(name);
+    public String deleteStudentTeacherPair(String teacher) {
 
-        studentTeacherPairDb.remove(name);
+        if (!studentTeacherPairDb.containsKey(teacher)) return "";
+
+        List<String> students = studentTeacherPairDb.get(teacher);
+
+        studentTeacherPairDb.remove(teacher);
 
         for (String student: students) {
             studentRepository.deleteStudent(student);
